refactor(integrations): extract AwsRoleRequirements
